Switch to generating certs on the fly since Python packaging was being a pain.
